Edit command will display window even if not modified
If an unsaved mail was edited and later q! 'ed, the window is not opened
again, resulting in loss of content. From now, if the action is "Edit" then
the same mail is displayed back in Outlook even if the user haven't modfied it

+++ vimlook.bas
@@ -134,7 +134,7 @@ CreateFile:
     ' Get File object of this file after user edits it
     newtimestamp = fso.GetFile(filename).DateLastModified
 
-    If oldtimestamp = newtimestamp Then
+    If oldtimestamp = newtimestamp And MailAction <> "Edit" Then
         ' User might have just :wq! 'ed
         fso.DeleteFile (filename)
         Exit Sub
